move createVrtFromFiles to raster_layer.py
